# Conflicts:
#	zapp.yaml

+++ .setup/config/setenv.sh
@@ -10,7 +10,7 @@ export LIB_DIR="$LOCAL_SCRIPTS_DIR/../lib"
 source "$LIB_DIR/utilities.sh"
 source "$LIB_DIR/colors.sh"
 source "$LIB_DIR/prerequisites.sh"
-
+export USER=$(printf '%s' "${USER:-${LOGNAME:-$(basename "$HOME")}}" | tr '[:lower:]' '[:upper:]')
 set +e
 # Load CICS/IMS credentials
 if [[ -f $HOME/.profile.bankz ]]; then
@@ -45,7 +45,7 @@ if [[ ! -f "$ENV_FILE" || "$ENV_FILE" -ot "$CONFIG_FILE" || "$ENV_FILE" -ot "${B
 # Global
 _BPXK_AUTOCVT=ON
 PYTHONUNBUFFERED=1
-ZOS_CURRENT_USER=$(printf '%s' "${USER:-${LOGNAME:-$(basename "$HOME")}}" | tr '[:lower:]' '[:upper:]')
+ZOS_CURRENT_USER=$(get_section_value 'global' 'zos_current_user')
 ZOS_ADMIN_USER=$(get_section_value 'global' 'zos_admin_user')
 ZOS_CA_LABEL=$(get_section_value 'global' 'zos_ca_label')
 ZOS_KEYRING=$(get_section_value 'global' 'zos_keyring')
@@ -146,7 +146,11 @@ IMS_DATASTORE=${IMS_DATASTORE:-$(get_section_value 'ims' 'datastore')}
 IMS_PLEX=${IMS_PLEX:-$(get_section_value 'ims' 'dfs_imsplex')}
 IMS_JAVA_CONF_PATH=${IMS_JAVA_CONF_PATH:-$(get_section_value 'ims' 'java_conf_path')}
 IMS_DFS_IMS_SSID=${IMS_DFS_IMS_SSID:-$(get_section_value 'ims' 'dfs_ims_ssid')}
+IMS_JAVA_FOLDER="${IMS_JAVA_FOLDER:-$(get_section_value 'ims' 'ims_java_dir')}"
 IMS_JAVA_HOME="${IMS_JAVA_HOME:-$(get_section_value 'ims' 'ims_java_home')}"
+IMS_IXVOLSER="${IMS_IXVOLSER:-$(get_section_value 'ims' 'ixvolser')}"
+IMS_IRLM_ENABLEMENT="${IMS_IRLM_ENABLEMENT:-$(get_section_value 'ims' 'irlm_enablement')}"
+IMS_DATABASE_LOCK_MANAGER_SERVER_NAME="${IMS_DATABASE_LOCK_MANAGER_SERVER_NAME:-$(get_section_value 'ims' 'database_lock_manager_server_name')}"
 
 # zconfig
 ZCONFIG_ZCB_HOME=$(get_section_value 'zconfig' 'zcb_home')
@@ -155,11 +159,12 @@ ZCONFIG_HOME="${ZCONFIG_HOME:-$(get_section_value 'zconfig' 'zconfig_home')}"
 # Debug
 DEBUG_HLQ=$(get_section_value 'debug' 'debug_hlq')
 DEBUG_TCPIP_HQL=$(get_section_value 'debug' 'tcpip_hlq')
+EQAPROF_CONF_DIR=$(get_section_value 'debug' 'eqaprof_conf_dir')
 
 # Db2
 DB2_HLQ="${DB2_HLQ:-$(get_section_value 'db2' 'db2_hlq')}"
 DB2_SSID="${DB2_SSID:-$(get_section_value 'db2' 'ssid')}"
-DB2_JAVA_HOME="${DB2_JAVA_HOME:-$(get_section_value 'db2' 'db2_java_home')}"
+DB2_JAVA_FOLDER="${DB2_JAVA_FOLDER:-$(get_section_value 'db2' 'db2_java_dir')}"
 
 # Zowe Configuration
 ZOWE_RSE_PROFILE=$(get_section_value 'zowe' 'rse_profile')
@@ -184,19 +189,20 @@ VARS_TO_CHECK=(
 
 error=0
 
-for var in "${VARS_TO_CHECK[@]}"; do
-  if [ -z "${!var}" ]; then
-    print_error "Error: variable '$var' is not set or is empty." >&2
-    error=1
-  fi
-done
-
-if [ "$error" -eq 1 ]; then
-  print_error "One or more variables are missing. Stopping script." >&2
-  rm -f "$ENV_FILE"
-  exit 1
+if [ "$(uname)" = "OS/390" ]; then
+    for var in "${VARS_TO_CHECK[@]}"; do
+      if [ -z "${!var}" ]; then
+        print_error "Error: variable '$var' is not set or is empty." >&2
+        error=1
+      fi
+    done
+    
+    if [ "$error" -eq 1 ]; then
+      print_error "One or more variables are missing. Stopping script." >&2
+      rm -f "$ENV_FILE"
+      exit 1
+    fi
+    print_info "All variables are properly set."
 fi
-
-print_info "All variables are properly set."
 
 export PATH=${PYTHON_HOME:-}/bin:$JAVA_HOME:/bin:$PATH
